Moves fetch_body!/1 to Downloader.fetch_body!/1
